requirements.txt
python-telegram-bot[job-queue]==21.6
openai==1.54.4
httpx==0.27.2
